Swift only has public VIP, make Keystone reflect that
This was the cause of the almost impossible to track "value is a
required option for Puppet::Parser::Resource::Param" error. We're
trying to use map_params on params that don't exist. From reading
previous code it seems that private/admin VIPs for Swift are not
present on purpose (e.g. Swift load balancer also expects only public
VIP), so i'm fixing it by using public VIP for all endpoints instead
of introducing new VIPs.