BZ #1123303 - Add manage_service settings to get puppet out of the way.
https://bugzilla.redhat.com/show_bug.cgi?id=1123303

This does not include rabbit, due to unique issues with that service.
This patch should make starting services more consistent for the included
services as pacemaker will have full control of starting and stopping the
services without the chance of puppet having already done so, which potentially
could cause confusion for pacemaker.